refactor: 优化video
